zubhub_backend/requirements.txt
amqp==2.6.1
argh==0.26.2
attrs==20.3.0
backcall==0.2.0
billiard==3.6.3.0
boto3==1.16.40
botocore==1.19.40
celery==4.4.0
certifi==2020.11.8
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==5.1
decorator==4.4.2
defusedxml==0.6.0
Django==2.2.7
django-allauth==0.40.0
django-appconf==1.0.4
django-capture-on-commit-callbacks==1.2.0
django-celery-email==3.0.0
django-celery-results==2.0.0
django-cors-headers==3.1.1
django-crispy-forms==1.8.0
django-debug-toolbar==2.0
django-js-asset==1.2.2
django-mptt==0.11.0
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
django-summernote==0.8.11.6
django-treebeard==4.5.1
djangorestframework==3.10.3
factory-boy==2.12.0
Faker==4.17.1
flower==0.9.5
gunicorn==19.9.0
humanize==3.1.0
idna==2.10
ipython==7.19.0
ipython-genutils==0.2.0
itypes==1.2.0
jedi==0.17.2
Jinja2==2.11.2
jmespath==0.10.0
kombu==4.6.11
MarkupSafe==1.1.1
more-itertools==8.6.0
oauthlib==3.1.0
openapi-codec==1.3.2
packaging==20.4
parso==0.7.1
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==6.2.1
pluggy==0.13.1
prometheus-client==0.8.0
prompt-toolkit==3.0.8
psycopg2-binary==2.8.3
ptyprocess==0.6.0
py==1.9.0
Pygments==2.7.2
PyJWT==1.7.1
pyparsing==2.4.7
pytest==5.3.5
pytest-django==3.8.0
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.4
PyYAML==5.1.2
requests==2.23.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
tornado==6.1
traitlets==5.0.5
twilio==6.51.1
uritemplate==3.0.1
urllib3==1.25.11
vine==1.3.0
watchdog==0.10.2
wcwidth==0.2.5
whitenoise==4.1.4
